+++ Assets/Scripts/Window.cs
@@ -1,9 +1,9 @@
 ﻿public class Window
 {
     private double[][] coordinates;
-    private int height;
+    private double height;
 
-    public Window(double[][] coordinates, int height)
+    public Window(double[][] coordinates, double height)
     {
         this.coordinates = coordinates;
         this.height = height;
@@ -14,7 +14,7 @@
         return coordinates;
     }
 
-    public int getHeight()
+    public double getHeight()
     {
         return height;
     }
@@ -24,7 +24,7 @@
         this.coordinates = coordinates;
     }
 
-    public void setHeight(int height)
+    public void setHeight(double height)
     {
         this.height = height;
     }
